==> code/0_preprocessing.ipynb <==
# Text pre-processing 
Normalization and other things 

#!pip install pyspellchecker
#import nltk
#nltk.download('stopwords')

import itertools
import multiprocessing as mp
import numpy as np
import nltk
import pandas as pd
import re
import sklearn
import string

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.linear_model import LogisticRegression 
from sklearn.model_selection import train_test_split
from sklearn.metrics import f1_score

from nltk import pos_tag
from nltk.tokenize import TweetTokenizer
from nltk.corpus import wordnet
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer 
import nltk

from textblob import TextBlob

## Basic pre-processing before removing contractions

# load data 
train_df = pd.read_csv('/Users/allisonwun-huikoh/Google Drive/___hertie/2019-fall/e1291-ml/e1291-group-project/news-data/us_news_test.csv')
train_df = train_df[train_df['headline'].notnull()]
train_df = train_df[train_df['description'].notnull()]
train_df = train_df[train_df['text'].notnull()]
valid_df = pd.read_csv('/Users/allisonwun-huikoh/Google Drive/___hertie/2019-fall/e1291-ml/e1291-group-project/news-data/us_news_valid.csv')
valid_df.columns = [col.replace('validate.', '') for col in valid_df.columns]
valid_df = valid_df[valid_df['headline'].notnull()]
valid_df = valid_df[valid_df['description'].notnull()]
valid_df = valid_df[valid_df['text'].notnull()]

# inspecting data 
valid_df.dtypes

 # lowercase 
train_df['headline'] = train_df['headline'].apply(lambda x: " ".join(x.lower() for x in x.split()))
valid_df['headline'] = valid_df['headline'].apply(lambda x: " ".join(x.lower() for x in x.split()))
train_df['description'] = train_df['description'].apply(lambda x: " ".join(x.lower() for x in x.split()))
valid_df['description'] = valid_df['description'].apply(lambda x: " ".join(x.lower() for x in x.split()))
train_df['text'] = train_df['text'].apply(lambda x: " ".join(x.lower() for x in x.split()))
valid_df['text'] = valid_df['text'].apply(lambda x: " ".join(x.lower() for x in x.split()))

# removing punctuation 
train_df['headline'] = train_df['headline'].str.replace('[^\w\s]','')
valid_df['headline'] = valid_df['headline'].str.replace('[^\w\s]','')
train_df['description'] = train_df['description'].str.replace('[^\w\s]','')
valid_df['description'] = valid_df['description'].str.replace('[^\w\s]','')
train_df['text'] = train_df['text'].str.replace('[^\w\s]','')
valid_df['text'] = valid_df['text'].str.replace('[^\w\s]','')

# removing stopwords 
stop = stopwords.words('english')
train_df['headline'] = train_df['headline'].apply(lambda x: " ".join(x for x in x.split() if x not in stop))
valid_df['headline'] = valid_df['headline'].apply(lambda x: " ".join(x for x in x.split() if x not in stop))
train_df['description'] = train_df['description'].apply(lambda x: " ".join(x for x in x.split() if x not in stop))
valid_df['description'] = valid_df['description'].apply(lambda x: " ".join(x for x in x.split() if x not in stop))
train_df['text'] = train_df['text'].apply(lambda x: " ".join(x for x in x.split() if x not in stop))
valid_df['text'] = valid_df['text'].apply(lambda x: " ".join(x for x in x.split() if x not in stop))

# common word removal 
freq1 = pd.Series(' '.join(train_df['headline']).split()).value_counts()[:10]
freq1 = list(freq1.index)
train_df['headline'] = train_df['headline'].apply(lambda x: " ".join(x for x in x.split() if x not in freq1))

freq2 = pd.Series(' '.join(valid_df['headline']).split()).value_counts()[:10]
freq2 = list(freq2.index)
valid_df['headline'] = valid_df['headline'].apply(lambda x: " ".join(x for x in x.split() if x not in freq2))

freq3 = pd.Series(' '.join(train_df['description']).split()).value_counts()[:10]
freq3 = list(freq3.index)
train_df['description'] = train_df['description'].apply(lambda x: " ".join(x for x in x.split() if x not in freq3))

freq4 = pd.Series(' '.join(valid_df['description']).split()).value_counts()[:10]
freq4 = list(freq4.index)
valid_df['description'] = valid_df['description'].apply(lambda x: " ".join(x for x in x.split() if x not in freq4))

freq5 = pd.Series(' '.join(train_df['text']).split()).value_counts()[:10]
freq5 = list(freq5.index)
train_df['text'] = train_df['text'].apply(lambda x: " ".join(x for x in x.split() if x not in freq5))

freq6 = pd.Series(' '.join(valid_df['text']).split()).value_counts()[:10]
freq6 = list(freq6.index)
valid_df['text'] = valid_df['text'].apply(lambda x: " ".join(x for x in x.split() if x not in freq6))


##### rare word removal
freq_1 = pd.Series(' '.join(train_df['headline']).split()).value_counts()[-10:]
freq_1 = list(freq_1.index)
train_df['headline'] = train_df['headline'].apply(lambda x: " ".join(x for x in x.split() if x not in freq_1))

freq_2 = pd.Series(' '.join(valid_df['headline']).split()).value_counts()[-10:]
freq_2 = list(freq_2.index)
valid_df['headline'] = valid_df['headline'].apply(lambda x: " ".join(x for x in x.split() if x not in freq_2))

freq_3 = pd.Series(' '.join(train_df['description']).split()).value_counts()[-10:]
freq_3 = list(freq_3.index)
train_df['description'] = train_df['description'].apply(lambda x: " ".join(x for x in x.split() if x not in freq_3))

freq_4 = pd.Series(' '.join(valid_df['description']).split()).value_counts()[-10:]
freq_4 = list(freq_4.index)
valid_df['description'] = valid_df['description'].apply(lambda x: " ".join(x for x in x.split() if x not in freq_4))

freq_5 = pd.Series(' '.join(train_df['text']).split()).value_counts()[-10:]
freq_5 = list(freq_5.index)
train_df['text'] = train_df['text'].apply(lambda x: " ".join(x for x in x.split() if x not in freq_5))

freq_6 = pd.Series(' '.join(valid_df['text']).split()).value_counts()[-10:]
freq_6 = list(freq_6.index)
valid_df['text'] = valid_df['text'].apply(lambda x: " ".join(x for x in x.split() if x not in freq_6))

# spelling correction 
train_df['headline'].apply(lambda x: str(TextBlob(x).correct()))
valid_df['headline'].apply(lambda x: str(TextBlob(x).correct()))
train_df['description'].apply(lambda x: str(TextBlob(x).correct()))
valid_df['description'].apply(lambda x: str(TextBlob(x).correct()))
train_df['text'].apply(lambda x: str(TextBlob(x).correct()))
valid_df['text'].apply(lambda x: str(TextBlob(x).correct()))

lemmatizer = WordNetLemmatizer()
tokenizer =  TweetTokenizer()
# spell = SpellChecker()

train_df.head()

def normalize_contractions(tokens):
	token_map = {
		"ain't" : 'is not',
		"aint" : 'is not',
		"aren't" : 'are not',
		"can't" : 'can not',
		"cannot" : 'can not',
		"could've" : 'could have',
		"couldn't" : 'could not',
		"couldnt" : 'could not',
		"didn't" : 'did not',
		"didnt" : 'did not',
		"doesn't" : 'does not',
		"doesnt" : 'does not',
		"don't" : 'do not',
		"gonna'" : 'going to',
		"gotta'" : 'got to',
		"hadn't" : 'had not',
		"hasn't" : 'has not',
		"haven't" : 'have not',
		"he'll" : 'he will',
		"he's" : 'he is',
		"he've" : 'he have',
		"how'd" : 'how did',
		"how'll" : 'how will',
		"how're" : 'how are',
		"how's" : 'how is',
		"i'd" : 'i would',
		"i'll" : 'i will',
		"i'm" : 'i am',
		"i'mm" : 'i am',
		"i've" : 'i have',
		"iäm" : 'i am',
		"isn't" : 'is not',
		"isnt" : 'is not',
		"it'd" : 'it would',
		"it'll" : 'it shall',
		"it's" : 'it is',
		"let's" : 'let us',
		"she'll" : 'she will',
		"she's" : 'she is',
		"should've" : 'should have',
		"shouldn't" : 'should not',
		"shouldnt" : 'should not',
		"that'll" : 'that will',
		"that's" : 'that is',
		"there's" : 'there is',
		"they'll" : 'they will',
		"they're" : 'they are',
		"theyre" : 'they are',
		"theyve" : 'they have',
		"wasn't" : 'was not',
		"we'll" : 'we will',
		"we'r" : 'we are',
		"we're" : 'we are',
		"we've" : 'we have',
		"weren't" : 'were not',
		"what's" : 'what is',
		"who'll" : 'who will',
		"who're" : 'who are',
		"who's" : 'who is',
		"why're" : 'why are',
		"won't" : 'will not',
		"would've" : 'would have',
		"wouldn't" : 'would not',
		"wouldnt" : 'would not',
		"you'll" : 'you will',
		"you're" : 'you are',
		"you've" : 'you have'
	}
	norm_tokens = []
	for t in tokens:
		if t in token_map.keys():
			for item in token_map[t].split():
				norm_tokens.append(item)
		else:
			for item in re.split('\W+', t.replace("'s","")):
				if item != '':
					norm_tokens.append(item)
	return(norm_tokens)

def get_wordnet_pos(treebank_tag):
	if treebank_tag.startswith('J'):
		return wordnet.ADJ
	elif treebank_tag.startswith('V'):
		return wordnet.VERB
	elif treebank_tag.startswith('N'):
		return wordnet.NOUN
	elif treebank_tag.startswith('R'):
		return wordnet.ADV
	else:
		return None
    
def lemma_tokenizer(text):
	text = text.replace('’',"'")
	tokens = normalize_contractions(text)


train_df['headline'] = train_df['headline'].apply(lemma_tokenizer)
